Put "Add user" dialog within activity to capture focus
Dialogs are not a part of view hierarchy so if the dialog loses focus, it cannot be reached by traversing.

When it is places within full-screen activity, then the focus is captured within activity and returns to the dialog. If the activity was embedded in the right side pane, like the dialog used to look, the dialog would be accessible by traversing but the focus would  be granted to the left pane of Settings app. Therefore this UI change is required.

Bug: 376815882
Test: atest UserSettingsTest
Flag: android.multiuser.place_add_user_dialog_within_activity
Change-Id: If9f280573d2b0da8deedcdeaeffc9db6160f27a0