We no longer need to monitor the shortcut changes on this screen
**Why remove the monitor?**
We started monitor the shortcut changes on this screen because we would
show the status of the shortcut as part of the summary of the
preference. However, the summary is too wordy, and we decided just show
the On/Off state of color correction/inversion. With that change, we
forgot to clean up unnecessaroy monitor on the shortcut status on this
screen.

Bug: 392190782
Test: manual
Flag: EXEMPT low risk clean up task
Change-Id: Iec6aca509c96237c4abb5c394ac5e7ffd78362eb