Added unit test for MobileNetworkPhoneNumberPreferenceController visibility
The new test checks whether the item is available when the
SubscriptionUtil.isSimHardwareVisible() method returns true, i.e. if
the config_show_sim_info boolean flag is true on non-test runtime.

The default visibility status for all tests is defined in the test
setup() method, as this will also simplify the addition of new checks
later on.

Bug: 392808943
Flag: EXEMPT test only
Test: atest MobileNetworkPhoneNumberPreferenceControllerTest

Change-Id: Icedef190a8140fd8164dd8035ae93735a666b9f0