Ensure certificate alias is always visible
If the user credentials dialog displays multiple elements (e.g. if the
certificate is used by a wifi configuration), the content size exceedes
the available space.

Currently, the alias text view is the only element which declares a
`layout_weight`. The LinearLayout failes to distribute the space
correctly, pushing it beyond the bounds.

Bug: 287533409
Flag: EXEMPT bugfix
Test: manual
Change-Id: Ieff1b3011080dabf823c7689a3781c6e52b9791d