Show per-displays pref groups after built-in
When the built-in display settings as well as CD settings are both
present in the fragment, show external display settings category after
built-in. This applies to the per-display fragment and the initial
display list fragment.

Stop showing per-display settings nested in a parent list, as this was
causing extra spacing and complicating the code.

Flag: com.android.settings.flags.display_topology_pane_in_display_list
Test: ExternalDisplayPreferenceFragmentTest
Test: manually check topology mode for single display, multiple displays, in both fragments
Test: manually check v1 UI for single display, multiple displays, in both fragments
Bug: b/352648432
Bug: b/396116157
Change-Id: I7fdf72d198988feb1e7559f96a54f7680cf5b8a6