Early check for SIM EID availability based on telephony or user admin
The SIM EID field was already not being shown if the
PackageManger.FEATURE_TELEPHONY_DATA feature flag was not set, per the
check in getIsAvailableAndUpdateEid().

The async availability update was introduced in the context of bug
304560734, because the UI could block until the EID value was ready.
The config_show_sim_info check was left in getAvailabilityStatus() as
that cannot block.

This change moves to getAvailabilityStatus() the availability check
based on the PackageManager feature flags (which are set on system
boot and not changed during runtime) and also based on whether the
user is admin or not, as with any other telephony related setting.
None of these two checks would lead to ANR state.

Bug: 395714454
Flag: EXEMPT bugfix
Test: atest SimEidPreferenceControllerTest

Change-Id: I3447091670345285132a2acaf45cd129c2b84c00