Show expanded display settings in list fragment
Stop distinguishing between the display list and per-display fragments.
Instead, always show all displays that are available. This is closer to
how the UI will eventually look, which is each display in a tab.

Setting an OnChangeListener for multiple preferences with the same key
did not work, even though the conflicting preferences had different
parent preferences. Only the listener of the top-most preference was
notified even when changing a different one. So this CL changes
PrefBasics so that it can set unique keys for otherwise identical
preferences.

Flag: com.android.settings.flags.display_topology_pane_in_display_list
Bug: b/352648432
Test: ExternalDisplayPreferenceFragmentTest
Test: set rotation of two overlay displays to different non-default values
Change-Id: Idabef1e03afd496bfc955927a2894f14f999a105