Adds supervision enablement flow.
This change adds an activity which will create the supervision profile
and trigger the activity to set a lock for the user when necessary. It
also updates the ConfirmSupervisionCredentialsActivity to launch
BiometricPrompt for this user.

Testing for SetupSupervisionActivity will be added in a follow-up.

Bug: 389712793
Flag: android.app.supervision.flags.enable_supervision_settings_screen
Test: atest SupervisionMainSwitchPreferenceTest

Change-Id: I53e9ddd7506913fa873f653d053ad2db8ecff0c0