Displays updater: force refresh on listener add
Refresh the preference every time the listener is added, not only for
the first callback registration. This is needed because when switching
fragments, the callback is unregistered by
ConnectedDeviceGroupController, so no updates are received in the child
fragments. This is handled differently by other device updaters.

UsbConnectionBroadcastReceiver invokes onReceive() in register() to
force a refresh. BluetoothDeviceUpdater requires the caller
to do a force refresh explicitly. This CL goes with the latter pattern.

Flag: com.android.settings.flags.display_topology_pane_in_display_list
Test: open CD settings fragment, remove all displays, then go back to CD entry fragment
Bug: b/398042811
Change-Id: Id139f227b1d59bde0174538b29b46fdd21ec0d24