Rename key to explicitly reference TELEPHONY_CALLING in user settings
The name KEY_ENABLE_TELEPHONY was too generic and could cause
confusion, because now calling/messaging/data features in Telephony
can be independently enabled depending on the supported capabilities
of the device.

This key refers to the entry to allow or disallow voice calling in the
user settings, so make it explicit that it is voice calling related.

Bug: 395714454
Test: mm && atest
Flag: EXEMPT minor refactor

Change-Id: I56ace031b6aed60c312f4c1298502e343a5c482a