Update ActionDisabledByAdvancedProtectionDialog
For UI, the dialog was migrated to SpaDialogWindowTypeActivity, which
has a Material3 theme and can support rendering over QuickSettings.

This change also adds logic around which message to show depending on
the received extras for feature ID and the dialog type.

Finally, the dialog will display the support button if the help Intent
URI exists and resolves to an activity. Tests around the support button
will be added in a follow up CL.

Bug: 389080341
Test: Visual test
Test: atest ActionDisabledByAdvancedProtectionDialogTest
Flag: EXEMPT bug fix
Change-Id: I0ed23cd44666a40ad54d755f628cd36ef06cc145