[Audiosharing] Refresh the dialog handling onNewIntent
If an instance of the activity already exists at the top of the current task, the system routes the intent to that instance through a call to its onNewIntent() method, rather than creating a new instance of the activity.

So we need manually trigger the handling logic in onNewIntent to refresh
the dialogs.

Test: atest
Flag: com.android.settingslib.flags.promote_audio_sharing_for_second_auto_connected_lea_device
Bug: 395786392
Change-Id: Iccb0ef649ee60cb97f7557943760f854d31ccc80