Settings: Do not show install info for Play Store app
Clicking the "App details" row in the info subpage for the Play Store
app results in a crash. It appears that this app is not considered a
mainline module and has an installer label, but this app is a special
case. Hide the row for the Play Store app.

Bug: 302093631
Test: atest com.android.settings.spa.app.appinfo.AppInstallerInfoPreferenceTest
Flag: EXEMPT bugfix
Change-Id: I291ee3e5dec2075381cb52d041982583cdd04323