Remove conditional controller additions from ResetDashboardFragment
The visibility of the controllers is already managed in each of their
getAvailabilityStatus() method, so there should be no need to
conditionally add them.

The case of the NetworkResetPreferenceController is also a bit
special. If it is not explicitly added here, it may be implicitly
added via ResetPreferenceController with an overridden visibility
logic, and it would then be unconditionally shown even if the
isMobileDataCapable() check fails.

Bug: 395714454
Test: mm && atest
Flag: EXEMPT bugfix
Change-Id: I5105eb44f2b933d41bbe6931e16a64eb58964457