Merge bundle global & type preference controllers.
This change allows the combined preference controller to be a central place to manage dependencies between the global switch and the individual type checkboxes. When the last individual checkbox is unchecked, the global switch turns off as well. When the global switch is flipped, the individual type preferences become visible or invisible.

Flag: android.service.notification.notification_classification
Bug: 380291070
Test: atest BundleCombinedPreferenceControllerTest, manual
Change-Id: I4e97d3d999c6114dd8c71100a0b0de1eb8f94e31