Change storage_summary string id to fix crash
ag/32322018 reverts storage_summary string for default locale but other
languages might be still using stale translations and cause
MissingFormatArgumentException crash.

As a quick fix, change the string id to make it as a new string. The
side-effect is that storage summary is not localized for a while.

Fix: 404116634
Flag: EXEMPT bugfix
Test: Verified with zh_TW locale
Change-Id: I5d7d6be2c1eb524622f1013120cb44403b1f6f29