Use isMobileDataCapable/isVoiceCapable helpers in network/eid
Disabling telephony support may be done either with the
config_show_sim_info config option, or based on whether
TelephonyManager reports data capabilities. The isMobileDataCapable()
helper method ensures both ways are supported.

Additionally, the logic is also updated to explicitly check for voice
support, so that disabling this telephony element happens only when
both mobile data and voice are unsupported.

The unit tests are updated to consider both no-data and no-voice
cases, and a helper method is introduced to ensure a valid EID value
is reported, as the EID controller won't be shown if the EID is empty.

The unit tests require runBlocking() because controller.refreshData()
is a suspend function.

Bug: 395714454
Test: mm && atest MobileNetworkEidPreferenceControllerTest
Flag: EXEMPT bugfix
Change-Id: Ib773001307b4bbafe2675c0a2c588556bfdcd012