Settings: change location settings description string
Updating the Location footer string so the description would implicitly
include both TELEPHONY and MESSAGING formats when they are available.
This change is done so that we can use this string in devices where only
one of TELEPHONY or MESSAGING is available, such as the AL device where
only MESSAGING could be available.

Bug: 397454540
Flag: EXEMPT bugfix
Change-Id: I864e498995e621bbf147187a38725b331867f7b5