Allow admins to remove their user as long as they are not the main user
Before this change, if a user was an admin, they would not be able to delete themselves from the device. Before introduction of multiple users, it meant that only main user could not remove themselves, but with multiple admins becoming a possible configuration, this has introduced an unexpected complication.

This change limits self-removal to only be disabled for main user instead of every admin.

Bug: 382007823
Test: atest UserSettingsTest
Flag: EXEMPT trivial change
Change-Id: Idc28a5fd919e874faba3062e20b698d1665da1bc