Fix a flaky test in EthernetInterfaceTracker
In EtherentDialogTest, we are creating an ethernet interface and that is
interfering with the interface count here. Since we are checking for the
presence of the interface with the id, checking the count isn't
necessary and could cause issues in the future if some other test keeps
adding new interfaces.

Removed the check for interface count to avoid this issue.

Flag: com.android.settings.connectivity.ethernet_settings

Test: Manual testing

Change-Id: I202112a10f1a57c7bc235f115c2c6a1eced8e6d4