Add focus state UI for screen flash color selector
When focusing on a screen flash option in the screen flash color
selector dialog, there's no obvious UI indicates which color is focused
now and makes it hard to use keyboard navigation to select the color.
Adds an alpha circle background to represent the focused state of the
color.

Flag: EXEMPT bugfix
Bug: 407944891
Test: manually test with keyboard
Change-Id: I2c954db3ffc40f698582f6aa83b5567b2bf275b2