Support "add new network" when WIFI_ENTRY_KEY is not specified in WifiDialog2
Currently, if someone start activity with
com.android.settings.WIFI_DIALOG intent, and pass in empty
KEY_CHOSEN_WIFIENTRY_KEY, it will create the dialog using the deprecated
one: WifiDialog and not working well (although no one is doing that).
This CL changes to use WifiDialog2 to show the configuration dialog for
new network. This is also a preparation for the new "add network" flow
we are going to add from Quick Settings.

UX mock: go/al-wifi-config-ux

Bug: 408065021
Flag: EXEMPT bugfix
Test: Manual test
  atest SettingsUnitTests:AddNetworkFragmentTest
  atest WifiConfigController2Test \
        WifiDialogActivityTest \
	WifiDialog2Test
Change-Id: I1e3e941c975e9ff121c448571272372a084f0535