connected display: merge injector classes into one
Merge the three injector classes into a single class, rewriting the
two older ones in Kotlin for a more compact implementation.

The injector classes didn't seem to be divided in a meaningful
manner, and I somewhat regret creating a new one only for the topology
pane. The first two classes are in a hierarchy together, but the base
class was never created as itself, so these did not need to be separate
either.

Having the classes separate was not a problem when they were mostly only
used in one file each (DisplayTopologyPreference and
ExternalDisplayPreferenceFragment) but I will soon need to pass an
injector to DisplayBlock so having Injectors that belonged only to
a single class each will probably be confusing. Hence this CL.

Bug: b/397231553
Test: atest ExternalDisplayUpdaterTest.java
Test: atest ExternalDisplayTestBase.java
Test: atest ResolutionPreferenceFragmentTest.java
Test: atest TopologyClampTest.kt
Test: atest TopologyScaleTest.kt
Test: atest ExternalDisplayPreferenceFragmentTest.java
Test: atest DisplayTopologyPreferenceTest.kt
Flag: EXEMPT refactor
Change-Id: I7ceb29644e3d07c852da6e9cfd2a5b41ee3b482a