Explicitly set screen lock to UNSPECIFIED in tests.
Initially the test was not resetting the screen lock that was set during the run. This caused this screen lock to be applied in all the tests that ran afterwards.

With this change the test that sets the screen lock now resets in the tearDown. And the test that requires screen lock to be UNSPECIFIED now explicitly sets it to the expected value.

Bug: 409853547
Test: atest UserSettingsTest
Flag: android.multiuser.require_pin_before_user_deletion
Change-Id: I983d8d1385624aa6508447eea1e6f5dc9d62f130