Create new state `SourceLostState`.
Previously when a source is lost, we remove the preference directly from the controller. Now we set the state to `SourceLostState` and let the `AudioStreamStateHandler` handle the state transition and preference removal. This is to get ready for adding target scanning functionality.

Flag: com.android.settingslib.flags.enable_le_audio_sharing
Test: atest
Bug: 395978182
Change-Id: I69c6eb1255682f66215626db21c64981f4ca6581