Trigger haptics preview after setting has been applied
An issue has been reported that there is no haptic preview when
toggling it from the vibration settings panel. This revealed a
larger problem, as the preference change API will be called before
the setting is applied, which causes the preview to use the
previous configuration. This CL changes the haptic preview
triggering to be asynchronous, so it is executed after the
preference has been applied.

Bug: 410756621
Test: Manual, robotest
Flag: EXEMPT bugfix
NO_IFTTT=No changes necessary

Change-Id: I64b2ba4153ddb580f05c4ece3b60a6c80545b084