Topology Pane: narrow padding between blocks
Before this patch, the padding between display blocks was so wide that
the potential highlight border of all displays would not overlap with
other potential border positions.

This patch makes the padding between blocks narrower so that the
highlight border consumes a bigger proportion of the padding space.

When mirroring a potentially-non-Bitmap wallpaper (see b/397231553),
we need to use a SurfaceView, and cannot use View.background to show
the wallpaper. This CL prepares to use a SurfaceView by making the
wallpaper image its own separate View (mWallpaperView) rather than a
single layer in a LayerDrawable. The rest of the layers are separate
Views too.

Test: manual check with 3 display blocks in the pane
Test: DisplayTopologyPreferenceTest
Bug: b/397231553
Flag: com.android.settings.flags.display_topology_pane_in_display_list
Change-Id: Ib57bca55fb56448aab3fe1eb022ae5518c3163c8