CD topology: block shows non-image wallpaper
Mirror the wallpaper surface using an API that is not specific to the
wallpaper type.

This has a problem in that if the wallpaper is not visible because of a
fullscreen app covering it, the wallpaper will not be rendered in the
display block. This is documented in b/412755041 and will be addressed
in the next CL.

Flag: com.android.settings.flags.display_topology_pane_in_display_list
Test: set emoji wallpaper on main display and add two other displays, and verify pane refreshes after a change
Test: DisplayBlockTest
Bug: b/397231553
Bug: b/412755041
Bug: b/416249965
Change-Id: I1548872c74822d7930682cdf4b605cbc1e44ab96