CD topology: ensure wallpaper is visible
If a fullscreen app is shown on a display, its wallpaper won't be
visible and we cannot mirror it. This forces the wallpaper to be
visible by adding a View to the window manager which shows the
wallpaper which would be underneath it.

Bug: b/412755041
Bug: b/415345340
Flag: com.android.settings.flags.display_topology_pane_in_display_list
Test: DisplayTopologyPreferenceTest
Change-Id: Iae52514e254939abc208827cee7b530a8f915c04