This change implements the Delete PIN flow
Introduces a dialog that will be shown when Delete PIN on
the Manage PIN screen is clicked. If users other than the current user
are supervised on the device, it will notify the user that they need to
disable supervision on those profiles first. If not, confirming the
dialog will disable supervision on the current user, delete recovery
info, and delete the supervising profle from the device.

Bug: 406082832
Flag: android.app.supervision.flags.enable_supervision_settings_screen
Test: manual

Change-Id: I9a71e6d98524ec2cd9795939357675c712dd1d60