UserSettingsTest: do not assume that current user is user 0
This is a preparatory CL before introducing new logic related
to handling headless system user under User section of Settings.

Previously tests were assuming that the current user is the user
with id 0, which is not a correct assumption on the devices that
have headless system user.

ShadowProcess is used to inject id of current user (code mostly
rely on UserHandle.myUserId() that derives value from process UID.

Another change is to consistently use mocked UserManager for all test
setups except for supportsMultipleUsers check. This check is implemented
as a static method, so this would remain the only function of
ShadowUserManager in test.

Test: atest com.android.settings.users.UserSettingsTest
Flag: TEST_ONLY
Bug: 394218826
Change-Id: I803c469f75011c056d6d368c7bdf6f3c47ac8425