Adding Settings items for Navbar customizable order
Defines options to change the order of the buttons on the Navbar.
This also migrates the Navbar's settings page to catalyst,
as it would've needed to be done anyway. This notably excludes the assistant setting present on the page, as to our understanding it's not intended to be used anymore, and is usually disabled.

NO_IFTTT= Change to SystemNavigationGestureSettings does not conflict with SystemNavigationGestureScreen

Flag: android.view.accessibility.navbar_flip_order_option
Bug: 410601652
Test: launch page manually
Test: atest com.android.settings.gestures
Change-Id: I2cf5f14f31ac8c22f2620412a912a8d43438a381