Revert^2 "usb: Initialize restricted radio button for USB settings"
This reverts commit 24a414abffe2061b7b542b645e07014c3f50886a.

Reason for revert: The initial revert was done because of bugs b/417967530 and b/418075552. However the investigation (thanks to rubinxu@google.com) shows that the reverted CL is not the culprit. That's why I am reverting the revert in order to keep the original submit https://googleplex-android-review.git.corp.google.com/c/platform/packages/apps/Settings/+/33507289.

Change-Id: I64ac89d89409012668ebdfb6d30260ef4d0a1420