Include LOCK_SCREEN_REDACTION in secondary user SuW "Anything else?"
The LOCK_SCREEN_REDACTION (notification on lock screen) is currently not
included in the suggested actions (aka "Anything else?") for secondary
user Setup Wizard (SuW), while it is included for the main user SuW.
This restriction was added in commit
dd05ab7600f79a6e77b6b5ed3916823d5985df72 ("Changes
SetupRedactionInterstitial to be an optional step") (CL:894059), and the
comments in the associated bug suggest that it's because lock screen was
not considered to show for secondary users. However it is currently
available for secondary users, and the corresponding "Notifications on
lock screen" page on Settings app is available for secondary users as
well, so the restriction is no longer necessary.

Bug: 418964276
Bug: 23893721 (comment #20)
Test: Check that the lock screen notification tile shows on the
  "Anything else" screen at the end of setup wizard for a secondary
  user. Check that the selection on the SuW page is preserved in the
  "Notifications on lock screen" Settings page after the setup.
Flag: EXEMPT bug fix

Change-Id: Ic74bbee2195f3d7c752d797581c4f91d856a1165