Add label for owner info EditText view in a dialog
Since the EditText uses the "hint" attribute
as an example of text that could be input into
the EditText, we cannot use contentDescription
(hints override it). Therefore, we must use
the title of the dialog to provide the label for
the EditText.

Fixes: 418746868
Test: enable talkback; Settings > Display > Lock screen >
Add text on lock screen > swipe right, then swipe left
to refocus on EditText; hear label being read for the
EditText box
Flag: EXEMPT bugfix

Change-Id: I4c63308be0ecc99b41430d9377e04fcfc718fea2