Cancel auth prompt after configuration change.
This is the easiest way to handle configuration changes. The magic that
androidx BiometricPrompt has can't be used here. The BiometricPrompt we use
advises to use setRetainIntance for fragment, which is deprecated, and
I'm also not sure if it's even appropriate for this fragment.
Alternatives are possible, but would be very complicated, as we'd need
to transition to some sort of ViewModel/LiveData system, which just
isn't needed for this sort of thing that isn't at all a problem for the
user.

Test: atest SettingsRoboTests:com.android.settings.display.ScreenTimeoutSettingsTest
Test: also manually tested
Bug: 348149031
Flag: EXEMPT bug fix
Change-Id: I697934e297604611ed3ab206cc4bd7e26663c922