Refactor DisplayTopologyPreference
With DisplayTopologyPane as a Preference, it will be included in 1
PreferenceScreen along with other Display settings (resolution,
rotation, etc.), in the same RecyclerView (important detail for later).

As we're trying to add tabbed display settings, this will cause list of
Preference to be changing. There are 2 approaches tried below:
1. Replacing / recreating fragment
This will surely cause DisplayTopologyPane to be recreated and caused it
to flicker (since it has SurfaceView)
2. Alternative: Hide/show the other display settings preference
As DisplayTopologyPreference and other prefs are in the same RecyclerView,
any change to an item will re-render the other items, in this case
DisplayTopologyPreference#onBindViewHolder will be called again
(which caused the flicker issue).
This happens regardless of PreferenceGroup as everything will eventually
get flattened into a single list
For example,
RecyclerView
|- displayTopologyPreference
|- resolutionPreference
|- rotationPreference

Therefore, the idea here would be to have a separated View / Fragment
for DisplayTopology panel. So, even if the other prefs get changed, it
will not get redrawn.

To minimize the duplicating when creating a new
DisplayTopology(Preference)View, most logics are extracted out from
DisplayTopologyPreference to a Controller to then be reused by View.

Bug: 409354332
Test: atest DisplayTopologyPreferenceTest
Flag: EXEMPT refactoring

Change-Id: Iba57ca7086a2c60dde3cbfe1ab0341ff4f6b5cd3