Observe use_primary_keys in mouse keys main page
When the use_primary_keys is true, the How to Use Primary Keys image
instruction and How to use a numeric keyboard image instruction
are both displayed.

When it's set to false, the How to use Primary Keys instruction
should be hidden.

Observing the use_primary_keys settings to update the image
instuctions as well as the toggle setting.

screen cast is added in the bug comments.

Bug: 419635632
Test: manually verified on device
Test: atest MouseKeysResetControllerTest
Flag: com.android.server.accessibility.enable_mouse_key_enhancement
Change-Id: Ia094e091794f45baa0d41d39460290c6b6a94de0