Fix sound work profile
In order to apply the new UI style, we added a PreferenceCategory for
each preference. However, there was no relevant child preference in the
new PreferenceCategory, which caused the preference of sound work
profile to disappear after enabling work profile. This change adds the
missing preference to PreferenceCategory.

Fix: 423494599
Test: manual test with testdpc
Flag: EXEMPT bugfix
Change-Id: Ib420afb992fee1e33607a2e9a9b2d46763b00415