Apply letterboxing strategy to mirroring displays
Due to varying aspect ratio of displays, DEFAULT_DISPLAY can't be
exactly mirrored. As the real display is also applying letterboxing,
it might be better to apply the same strategy to the mirror preview

Screenshots:
Light mode: http://shortn/_pH1krcawZr
Dark mode: http://shortn/_oTG33Kp4Xk
2 external displays: http://shortn/_kyWrbzsode

The 2 external displays one is a bit broken as the non-letterboxed area
is translucent. Got a prototype fix available, but it's not so
straightforward. As the system doesn't allow having more than 1 external
displays for now, we can go with this for the current release, and fix
the issue after

Bug: 425355749
Test: atest DisplayBlockTest
Test: atest DisplayTopologyPreferenceTest
Flag: com.android.settings.flags.show_stacked_mirroring_display_connected_display_setting
Change-Id: I577eead90a980716390f4b77c04fba47fcb6aabb