Inline observeAsCallback from settingslib
The `observeAsCallback` extension function is moved from
`com.android.settingslib.spa.livedata` into a new local file
`com.android.settings.spa.livedata.LiveDataExt.kt`.

This change allows for the removal of a dependency on the corresponding
settingslib module. Since the function is already deprecated in favor of
using Flow, inlining it for the few remaining usages simplifies the
dependency graph.

The necessary `runtime-livedata` and `lifecycle-livedata-ktx`
dependencies are added to `Android.bp`, and the call sites in
`NfcTagAppsSettings` and `AppNotificationsListModel` are updated to use
the new local function.

Bug: 321163306
Flag: EXEMPT clean up
Test: m Settings
Change-Id: Ic596163c0bbd5e93378732e61fb47a7a46bd9638