Run ChooseLock* for supervising profile as current user
Modifies Settings activities for choosing a lock to be run for the
supervising profile (which is a "parentless profile" -- only one exists
on the device, and it is considered a valid profile for any user) within
the context of the current user.

This allows us to avoid the significant storage overhead of
pre-installing settings on the supervising profile user, as well as
avoid the extra latency of starting the activity in a separate user.

Bug: 417999401
Bug: 419810809
Flag: android.multiuser.allow_supervising_profile
Test: atest SetupSupervisionActivityTest
Test: atest SupervisionChangePinPreferenceTest
Test: atest SupervisionPinRecoveryActivityTest

Change-Id: Ia8b13041aac9c9dddffe06fc06f543461734292d